% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adacov.R
\name{adacov}
\alias{adacov}
\title{Adaptive Thresholding for Sparse Covariance Matrix Estimation}
\usage{
adacov(X, delta = 2, eta = 4)
}
\arguments{
\item{X}{an n by p data matrix.}

\item{delta}{a numeric value. A tuning parameter.
The default value of delta is 2, and is taken from Cai and Liu (2011).
Also, it can be chosen empirically through cross-validation.}

\item{eta}{a numeric value. A shrinkage parameter. The greater value
  shrink sparse values more.}
}
\value{
Test results
}
\description{
A templete function for adaptive thresholding for covariance matrix.
This method varies by how theta was defined, and what funtion form for
adaptive thresholding was used.
}
\references{
will be added
}

