% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diag2}
\alias{diag2}
\title{A convenient version of diag}
\usage{
diag2(x = 1, nrow, ncol, off = 0)
}
\arguments{
\item{x}{}

\item{nrow}{}

\item{ncol}{}

\item{off}{an integer. location of off-diagonal}
}
\value{
Test results
}
\description{
text
}
\details{
...
}
\examples{
\dontrun{
}
}
\author{
Dongjun You
}

