% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genCLX14}
\alias{genCLX14}
\title{Generate simulation data for CLX14}
\usage{
genCLX14(n, p, m = c("m1", "m2", "power"), muval = c("muval1", "muval2"),
  model = c("model1", "model2", "model3"))
}
\arguments{
\item{n}{nobs of X1 and X2}

\item{p}{number of variables}

\item{m}{a character value. Two values of the number of non-zero mu1.
For 'm1', floor(0.05p) of mu1 will be set to non-zeros.
floor(sqrt(p)) for 'm2'. 0 for 'power'.}

\item{muval}{a character value. The magnitude of the elements of mu1.
If 'muval1', pm sqrt(log(p)/n). 'muval2' for
rnuif(-sqrt(8*log(p)/n), sqrt(8*log(p)/n)).}

\item{model}{a character value. A model generating covariance matrix.}
}
\value{
Test results
#'
}
\description{
TEST
}
\examples{
\dontrun{
res <- genCLX14(n=50, p=50)
X1 <- res[["X1"]]
X2 <- res[["X2"]]
hotelhd(X1, X2, method="H")
hotelhd(X1, X2, method="D")
hotelhd(X1, X2, method="BS")
hotelhd(X1, X2, method="CQ")
}
}
\author{
Dongjun You
}
\references{
will be added
}

