% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genCQ10}
\alias{genCQ10}
\title{Simulation data for CQ10}
\usage{
genCQ10(n, p, perc, dist = c("gamma", "norm"), dependence = c("two",
  "full"))
}
\arguments{
\item{n}{nobs of X1 and X2}

\item{p}{number of variables}

\item{perc}{percentages}

\item{dist}{gamma or norm}

\item{dependence}{two or full}
}
\value{
Test results
#'
}
\description{
TEST
}
\examples{
\dontrun{
res <- genCQ10(n=124, p=500, perc=50)
X1 <- res[["X1"]]
X2 <- res[["X2"]]
hotelhd(X1, X2, method="H")
hotelhd(X1, X2, method="D")
hotelhd(X1, X2, method="BS")
hotelhd(X1, X2, method="CQ")
}
}
\author{
Dongjun You
}
\references{
will be added
}

