% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hotelhd.R
\name{hotelhd}
\alias{hotelhd}
\title{Two Sample Inference for High Dimensional Mean}
\usage{
hotelhd(X1, X2, method = c("H", "D", "BS", "CQ", "CLX", "Z", "M"),
  alpha = 0.05, control = hotelhd_control(method = method))
}
\arguments{
\item{X1,X2}{matrix. n1(,n2) by p data matrix. Missing values (including NaN) will
be omitted from the calculations.}

\item{method}{character. The method to be used.
See 'Details'.}

\item{alpha}{a numeric value. The type 1 error.}

\item{control}{a list with control parameters as returned by \code{hotelhd_ontrol}.}
}
\value{
Test results
}
\description{
\code{hotelhd} is used to test two sample inference problem for high dimensional
  data.
}
\details{
Data 'X1' and 'X2' only accept matrix as arguments. All the methods
implemented in this package assume continuous variables as inputs.

Method "H" is for Hotelling's T2 test.

Method "D" is for Dempster's (1958) non-exact test.

Method "BS" is

Method "CQ" is

Method "CLX" is using CLIME to estimate Omega. The current version
  uses the default options of 'fastclime' function of
  'fastclime' package.
  Whether the null is rejected or not is only included in the result,
  because of the form of asymptotic distribution of the test statistics.
  A function 'rejected' for hypothesis test will be returned, which
  needs an argument 'alpha' to test Whether the null is rejected or not.

Method "Z" is blockwise bootstrap

Method "M" is generalized bootstrap
}
\references{
Dempster, A. P. (1958). A high dimensional two sample significance test. \emph{The Annals of Mathematical Statistics}, 995-1010.
}

