% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diag2.R
\name{diag2}
\alias{diag2}
\title{A convenient version of diag}
\usage{
diag2(x = 1, nrow, ncol, off = 0)
}
\arguments{
\item{x}{a value to be filled}

\item{nrow}{nrow}

\item{ncol}{ncol}

\item{off}{an integer. location of off-diagonal}
}
\value{
Test results
}
\description{
text
}
\author{
Dongjun You
}

