% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genCQ10}
\alias{genCQ10}
\title{Simulation data for CQ10}
\usage{
genCQ10(n1, n2, p, rho, eta, dist = c("gamma", "norm"))
}
\arguments{
\item{n1}{nobs of X1}

\item{n2}{nobs of X2}

\item{p}{number of variables}

\item{dist}{gamma or norm}
}
\value{
Test results
#'
}
\description{
TEST
}
\examples{
\dontrun{
res <- genBS96(n1=25, n2=20, p=40, rho=0.5, eta=0, dist="norm")
res[["phi"]]
res[["lambda"]]
d1 <- res[["d1"]]
d2 <- res[["d2"]]
hotelhd(d1, d2, method="H")
hotelhd(d1, d2, method="D")
hotelhd(d1, d2, method="BS")
hotelhd(d1, d2, method="CQ")
}
}
\author{
Dongjun You
}
\references{
will be added
}

