% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hotelhd}
\alias{hotelhd}
\title{Hotelling's Test with High Dimensional Data}
\usage{
hotelhd(d1, d2, type = c("T2", "nonexact", "BS"), na.rm = TRUE)
}
\arguments{
\item{d1}{a dataset. \code{as.matrix} will be applied.}

\item{d2}{a dataset. \code{as.matrix} will be applied.}
}
\value{
Test results
}
\description{
TEST
}
\details{
Users should take care to ensure that the two dataset have
the same variables and the same order of the variables
}
\examples{
\dontrun{
}
}
\author{
Dongjun You
}
\references{
will be added
}

