% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hotelhd}
\alias{hotelhd}
\title{Hotelling's Test with High Dimensional Data}
\usage{
hotelhd(X1, X2, na.rm = TRUE, method = c("H", "D", "BS", "CQ", "CLX", "Z"),
  C = 10, omega = c("clime", "ada"), R = 500, block = 1, alpha = 0.05)
}
\arguments{
\item{X1}{a n1 by p data matrix. See Details.}

\item{X2}{a n2 by p data matrix.}

\item{na.rm}{logical value of TRUE or FALSE.
option for calculating means of p varuables}

\item{method}{method to be used. See 'Details'.}

\item{C}{a constant to decide lambda_n for 'CLX', See Cai et al. (2011).}

\item{omega}{a character value. the estimation of Omega matrix for 'CLX'.
'clime' for clime, 'ada' for Adaptive Thresholding.
The definition of 'ada' follows CLX14's approach for two-sample test.}

\item{R}{a numeric value. The number of bootstrap statistics for 'Z'.}

\item{block}{a numeric value. A block size for blockwize multiplier
bootstrap in the method 'Z'. The size should be smaller
than min(n1, n2). The default value is 1.}

\item{alpha}{a numeric value. The type 1 error.}
}
\value{
Test results
}
\description{
Users should take care to ensure that the two dataset have
the same variables and the same order of the variables
}
\details{
Data 'X1' and 'X2' only accept matrix as arguments. All the methods
implemented in this package assume continuous variables as inputs.

Method "H" is for Hotelling's T2 test.

Method "D" is for Dempster's (1958) non-exact test.

Method "BS" is

Method "CQ" is

Method "CLX" is using CLIME to estimate Omega. The current version
  uses the default options of 'fastclime' function of
  'fastclime' package.
  Whether the null is rejected or not is only included in the result,
  because of the form of asymptotic distribution of the test statistics.
  A function 'rejected' for hypothesis test will be returned, which
  needs an argument 'alpha' to test Whether the null is rejected or not.
}
\references{
Dempster, A. P. (1958). A high dimensional two sample significance test. \emph{The Annals of Mathematical Statistics}, 995-1010.
}

