% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hotelhd}
\alias{hotelhd}
\title{Hotelling's Test with High Dimensional Data}
\usage{
hotelhd(X1, X2, na.rm = TRUE, method = c("H", "D", "BS", "CQ", "CLX"))
}
\arguments{
\item{X1}{a n1 by p data matrix. Only a matrix can be used.}

\item{X2}{a n2 by p data matrix.}

\item{na.rm}{logical value of TRUE or FALSE.
option for calculating means of p varuables}

\item{method}{method to be used. See 'Details'.}
}
\value{
Test results
}
\description{
Users should take care to ensure that the two dataset have
the same variables and the same order of the variables
}
\details{
Method "H" is ...

Method "D" is

Method "BS" is

Method "CQ" is

Method "CLX" is
}
\author{
Dongjun You
}
\references{
will be added
}

