% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hotelhd.R
\name{hotelhd}
\alias{hotelhd}
\title{Two Sample Inference for High Dimensional Mean}
\usage{
hotelhd(X1, X2, method = c("H", "D", "BS", "CQ", "CLX", "Z", "M"),
  alpha = 0.05, control = hotelhd_control(method = method))
}
\arguments{
\item{X1,X2}{matrix. n1(,n2) by p data matrix. Missing values (including NaN) will
be omitted from the calculations.
Data 'X1' and 'X2' only accept matrix as arguments to prevent confusion.
All the methods
implemented in this package assume continuous variables as inputs.}

\item{method}{character. The method to be used.
See 'Details'.}

\item{alpha}{numeric. The type 1 error.}

\item{control}{list. Control parameters returned by \code{hotelhd_control}.
  Users NEED to check the function \code{\link{hotelhd_control}} when applying
  maximum type tests (i.e., "CLX", "Z", "M").}
}
\value{
Test results.
}
\description{
\code{hotelhd} is used to test two sample inference problem for high dimensional
  data.
}
\details{
\describe{
  \item{H}{Hotelling's T-squared test.
    Implemented for convenient comparisons with other methods.
    Users can find more sophisticated implementation of
    Hotelling's T-squared test in CRAN.}
  \item{D}{Dempster's (1958) non-exact test.}
  \item{BS}{Bai and Saranadasa's (1996) test.}
  \item{CQ}{Chen and Qin's (2010) test.}
  \item{CLX}{Cai, Liu and Xia's (2014) test.}
  \item{Z}{blockwise bootstrap}
  \item{M}{generalized bootstrap}
}
}
\references{
\itemize{
\item Bai, Zhidong D and Saranadasa, Hewa. (1996). Effect of high dimension: by an example of a two sample problem. \emph{Statistica Sinica}, 311-329.
\item Cai, T. Tony and Liu, Weidong and Xia, Yin. (2014). Two-sample test of high dimensional means under dependence. \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, 349-372.
\item Chen, Song Xi and Qin, Ying-Li. (2010). A two-sample test for high-dimensional data with applications to gene-set testing. \emph{Annals of Statistics}, 808-835.
\item Dempster, A. P. (1958). A high dimensional two sample significance test. \emph{The Annals of Mathematical Statistics}, 995-1010.
}
}

