% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hotelhd_control.R
\name{hotelhd_control}
\alias{hotelhd_control}
\title{Control Parameters for \code{hotelhd}}
\usage{
hotelhd_control(C = 10, omegaGiven = NULL, omegaHat = c("omega",
  "identity"), omegaEst = c("clime"), subForM = c("sub", "diag"),
  ndim = 4L, R = 500, block = 1)
}
\arguments{
\item{C}{numeric. to decide lambda_n for 'CLX', See Cai et al. (2011).}

\item{omegaGiven}{matrix; \code{NULL} or \eqn{p} by \eqn{p} matrix.
  When the default \code{NULL} is chosen, \eqn{\Omega} matrix is estimated;
  see \code{omegaHat} and \code{omegaEst} parameters.
  Optionally, users can provide an estimate of \eqn{\Omega}
  from an estimation method that this package does not offer.}

\item{omegaHat}{character, 'omega' or 'identity'.
  A candidate estimate of \eqn{\Omega} matrix for the maximum type methods.
  When the default of 'omega' is used, \eqn{\Omega} is estimated.
  Users can choose estimation method by \code{omegaEst} parameter.
  Optionally,
  'identity' can be choosen to set \eqn{p} by \eqn{p} Identity matrix as
  an estimate of \eqn{\Omega} matrix.}

\item{omegaEst}{character, 'clime'.
  The estimation method for \eqn{\Omega} matrix.
  Currently, only 'clime' is implemented; see \code{\link[flare]{sugm}}.}

\item{ndim}{integer. The number of dimensions for generalized maximum
  type statistic, i.e., the method 'M'. The default is 4L.}

\item{R}{The number of bootstrap statistics for 'Z'.}

\item{block}{A block size for blockwize multiplier
  bootstrap in the method 'Z'. The size should be smaller
  than min(n1, n2). The default value is 1.}

\item{sub}{character; 'diag' or 'sub'.
Method about how to form \code{ndim} by \code{ndim} sub-matrix for
the method 'M'. The default 'diag' uses diagonals of \eqn{\Omega}
to improve efficiency. 'sub' extracts sub-matrix from \eqn{\Omega}.}
}
\description{
Various parameters that control setting of tests. Currently,
the maximum type tests (i.e., 'CLX', 'Z', 'M') only need control parameters.
}

