% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hotelhd_control.R
\name{hotelhd_control}
\alias{hotelhd_control}
\title{Control Parameters for \code{hotelhd}}
\usage{
hotelhd_control(method = c("H", "D", "BS", "CQ", "CLX", "Z", "M"), C = 10,
  omegaHat = c("omega", "identity"), omegaEst = c("clime", "ada"),
  omegaGiven = NULL, subForM = c("sub", "diag"), ndim = 4, R = 500,
  block = 1)
}
\arguments{
\item{method}{character. The method to be used.
Each test needs different paremeters. See 'Details'.}

\item{C}{a constant to decide lambda_n for 'CLX', See Cai et al. (2011).}

\item{omegaHat}{A candidate estimate of Omega matrix for the two
maximum
type metods, 'CLX' and 'Z'. When the default 'omega' is used,
Omega will be estimated, See 'omegaHat' argument. Optionally,
'identity' can be choosen to set p by p Identity matrix as
the estimate of Omega matrix.}

\item{omegaEst}{The estimation method for Omega matrix for 'CLX'.
'clime' for clime, 'ada' for Adaptive Thresholding.
The definition of 'ada' follows CLX14's approach for two-sample test.}

\item{omegaGiven}{If an omega was given, use it. This is an optional but
is included to improve speed of simulation study.
All the methods using 'clime' package will use the given omega,
so do not need to calculate individually.}

\item{subForM}{for the method 'M'. 'sub' extract sub-matrix from Omega.
'diag' use diagonals of Omega (to improve efficiency)}

\item{ndim}{The number of dimensions for generalized maximum type statistic.
Default value is 4.}

\item{R}{The number of bootstrap statistics for 'Z'.}

\item{block}{A block size for blockwize multiplier
  bootstrap in the method 'Z'. The size should be smaller
  than min(n1, n2). The default value is 1.}
}
\description{
Various parameters that control setting of tests.
}

